+++ mb/ch06/Token.hpp
@@ -6,7 +6,6 @@ public:
     Token(char);
     Token(char, double);
 
-
     char kind;
     double value;
 };
