+++ mb/ch06/TokenStream.hpp
@@ -1,21 +1,8 @@
 #pragma once
 
 #include <sstream>
+#include "ITokenStream.hpp"
 #include "Token.hpp"
-
-const char QUIT = 'k';
-const char END_OF_EXPR = ';';
-const char NUMBER = 'd';
-
-class ITokenStream
-{
-public:
-    virtual ~ITokenStream() {};
-    
-    virtual Token get() = 0;
-    virtual void putback(Token) = 0;
-    virtual void ignore(char) = 0;
-};
 
 class TokenStream : public ITokenStream
 {
